+++ src/pages/AddCard.tsx
@@ -1,7 +1,7 @@
 import CardNumbers from '../components/CardNumbers/CardNumbers';
-import ExpiredDate from '../components/ExpiredDate/ExpiredDates';
+import ExpiredDate from '../components/CardExpiredDate/CardExpiredDates';
 import CardOwnerName from '../components/CardOwnerName/CardOwnerName';
-import SecurityCode from '../components/SecurityCode/SecurityCode';
+import SecurityCode from '../components/CardSecurityCode/CardSecurityCode';
 import CardPassword from '../components/CardPassword/CardPassword';
 import Card from '../components/Card/Card';
 import Header from '../components/Header/Header';
